mediabank-qa-automation-metadata/pyproject.toml
[tool.poetry]
name = "automation_test_python"
version = "0.1.0"
description = "Template for backend automation test"
authors = ["QA"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.5"
pytest-bdd = "^5.0.0"
docker = "^6.0.1"
python-on-whales = "^0.59.0"
pytest-dotenv = "^0.5.2"
psycopg2-binary = "^2.9.5"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
    'ignore::DeprecationWarning:docker.*:'
]
log_cli = true
log_cli_format = "[%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
